+++ src/Commands/MigrateRefreshCommand.php
@@ -28,8 +28,16 @@ class MigrateRefreshCommand extends Command
     /**
      * Execute the console command.
      */
-    public function handle()
+    public function handle() : int
     {
+        $module = $this->argument('module');
+
+        if ($module && !$this->getModuleName()) {
+            $this->error("Module [$module] does not exists.");
+
+            return E_ERROR;
+        }
+
         $this->call('module:migrate-reset', [
             'module' => $this->getModuleName(),
             '--database' => $this->option('database'),
@@ -47,6 +55,8 @@ class MigrateRefreshCommand extends Command
                 'module' => $this->getModuleName(),
             ]);
         }
+
+        return 0;
     }
 
     /**
@@ -79,12 +89,12 @@ class MigrateRefreshCommand extends Command
     {
         $module = $this->argument('module');
 
-        $module = app('modules')->find($module);
-
-        if ($module === null) {
-            return $module;
+        if (!$module) {
+            return null;
         }
 
-        return $module->getStudlyName();
+        $module = app('modules')->find($module);
+
+        return $module ? $module->getStudlyName() : null;
     }
 }