+++ src/Commands/UnUseCommand.php
@@ -23,10 +23,12 @@ class UnUseCommand extends Command
     /**
      * Execute the console command.
      */
-    public function handle()
+    public function handle() : int
     {
         $this->laravel['modules']->forgetUsed();
 
         $this->info('Previous module used successfully forgotten.');
+
+        return 0;
     }
 }