+++ tests/Commands/EnableCommandTest.php
@@ -12,11 +12,13 @@ class EnableCommandTest extends BaseTestCase
     {
         parent::setUp();
         $this->artisan('module:make', ['name' => ['Blog']]);
+        $this->artisan('module:make', ['name' => ['Taxonomy']]);
     }
 
     public function tearDown(): void
     {
         $this->app[RepositoryInterface::class]->delete('Blog');
+        $this->app[RepositoryInterface::class]->delete('Taxonomy');
         parent::tearDown();
     }
 
@@ -27,8 +29,26 @@ class EnableCommandTest extends BaseTestCase
         $blogModule = $this->app[RepositoryInterface::class]->find('Blog');
         $blogModule->disable();
 
-        $this->artisan('module:enable', ['module' => 'Blog']);
+        $code = $this->artisan('module:enable', ['module' => 'Blog']);
 
         $this->assertTrue($blogModule->isEnabled());
+        $this->assertSame(0, $code);
+    }
+
+    /** @test */
+    public function it_enables_all_modules()
+    {
+        /** @var Module $blogModule */
+        $blogModule = $this->app[RepositoryInterface::class]->find('Blog');
+        $blogModule->disable();
+
+        /** @var Module $taxonomyModule */
+        $taxonomyModule = $this->app[RepositoryInterface::class]->find('Taxonomy');
+        $taxonomyModule->disable();
+
+        $code = $this->artisan('module:enable');
+
+        $this->assertTrue($blogModule->isEnabled() && $taxonomyModule->isEnabled());
+        $this->assertSame(0, $code);
     }
 }