+++ tests/Commands/ListenerMakeCommandTest.php
@@ -36,18 +36,19 @@ class ListenerMakeCommandTest extends BaseTestCase
     /** @test */
     public function it_generates_a_new_event_class()
     {
-        $this->artisan(
+        $code = $this->artisan(
             'module:make-listener',
             ['name' => 'NotifyUsersOfANewPost', 'module' => 'Blog', '--event' => 'UserWasCreated']
         );
 
         $this->assertTrue(is_file($this->modulePath . '/Listeners/NotifyUsersOfANewPost.php'));
+        $this->assertSame(0, $code);
     }
 
     /** @test */
     public function it_generated_correct_sync_event_with_content()
     {
-        $this->artisan(
+        $code = $this->artisan(
             'module:make-listener',
             ['name' => 'NotifyUsersOfANewPost', 'module' => 'Blog', '--event' => 'UserWasCreated']
         );
@@ -55,12 +56,27 @@ class ListenerMakeCommandTest extends BaseTestCase
         $file = $this->finder->get($this->modulePath . '/Listeners/NotifyUsersOfANewPost.php');
 
         $this->assertMatchesSnapshot($file);
+        $this->assertSame(0, $code);
+    }
+
+    /** @test */
+    public function it_generated_correct_sync_event_in_a_subdirectory_with_content()
+    {
+        $code = $this->artisan(
+            'module:make-listener',
+            ['name' => 'NotifyUsersOfANewPost', 'module' => 'Blog', '--event' => 'User/WasCreated']
+        );
+
+        $file = $this->finder->get($this->modulePath . '/Listeners/NotifyUsersOfANewPost.php');
+
+        $this->assertMatchesSnapshot($file);
+        $this->assertSame(0, $code);
     }
 
     /** @test */
     public function it_generated_correct_sync_duck_event_with_content()
     {
-        $this->artisan(
+        $code = $this->artisan(
             'module:make-listener',
             ['name' => 'NotifyUsersOfANewPost', 'module' => 'Blog']
         );
@@ -68,12 +84,13 @@ class ListenerMakeCommandTest extends BaseTestCase
         $file = $this->finder->get($this->modulePath . '/Listeners/NotifyUsersOfANewPost.php');
 
         $this->assertMatchesSnapshot($file);
+        $this->assertSame(0, $code);
     }
 
     /** @test */
     public function it_generated_correct_queued_event_with_content()
     {
-        $this->artisan(
+        $code = $this->artisan(
             'module:make-listener',
             ['name' => 'NotifyUsersOfANewPost', 'module' => 'Blog', '--event' => 'UserWasCreated', '--queued' => true]
         );
@@ -81,12 +98,27 @@ class ListenerMakeCommandTest extends BaseTestCase
         $file = $this->finder->get($this->modulePath . '/Listeners/NotifyUsersOfANewPost.php');
 
         $this->assertMatchesSnapshot($file);
+        $this->assertSame(0, $code);
+    }
+
+    /** @test */
+    public function it_generated_correct_queued_event_in_a_subdirectory_with_content()
+    {
+        $code = $this->artisan(
+            'module:make-listener',
+            ['name' => 'NotifyUsersOfANewPost', 'module' => 'Blog', '--event' => 'User/WasCreated', '--queued' => true]
+        );
+
+        $file = $this->finder->get($this->modulePath . '/Listeners/NotifyUsersOfANewPost.php');
+
+        $this->assertMatchesSnapshot($file);
+        $this->assertSame(0, $code);
     }
 
     /** @test */
     public function it_generated_correct_queued_duck_event_with_content()
     {
-        $this->artisan(
+        $code = $this->artisan(
             'module:make-listener',
             ['name' => 'NotifyUsersOfANewPost', 'module' => 'Blog', '--queued' => true]
         );
@@ -94,6 +126,7 @@ class ListenerMakeCommandTest extends BaseTestCase
         $file = $this->finder->get($this->modulePath . '/Listeners/NotifyUsersOfANewPost.php');
 
         $this->assertMatchesSnapshot($file);
+        $this->assertSame(0, $code);
     }
 
     /** @test */
@@ -101,7 +134,7 @@ class ListenerMakeCommandTest extends BaseTestCase
     {
         $this->app['config']->set('modules.paths.generator.listener.path', 'Events/Handlers');
 
-        $this->artisan(
+        $code = $this->artisan(
             'module:make-listener',
             ['name' => 'NotifyUsersOfANewPost', 'module' => 'Blog']
         );
@@ -109,6 +142,7 @@ class ListenerMakeCommandTest extends BaseTestCase
         $file = $this->finder->get($this->modulePath . '/Events/Handlers/NotifyUsersOfANewPost.php');
 
         $this->assertMatchesSnapshot($file);
+        $this->assertSame(0, $code);
     }
 
     /** @test */
@@ -116,7 +150,7 @@ class ListenerMakeCommandTest extends BaseTestCase
     {
         $this->app['config']->set('modules.paths.generator.listener.namespace', 'Events\\Handlers');
 
-        $this->artisan(
+        $code = $this->artisan(
             'module:make-listener',
             ['name' => 'NotifyUsersOfANewPost', 'module' => 'Blog']
         );
@@ -124,5 +158,6 @@ class ListenerMakeCommandTest extends BaseTestCase
         $file = $this->finder->get($this->modulePath . '/Listeners/NotifyUsersOfANewPost.php');
 
         $this->assertMatchesSnapshot($file);
+        $this->assertSame(0, $code);
     }
 }