+++ tests/Commands/RequestMakeCommandTest.php
@@ -35,19 +35,21 @@ class RequestMakeCommandTest extends BaseTestCase
     /** @test */
     public function it_generates_a_new_form_request_class()
     {
-        $this->artisan('module:make-request', ['name' => 'CreateBlogPostRequest', 'module' => 'Blog']);
+        $code = $this->artisan('module:make-request', ['name' => 'CreateBlogPostRequest', 'module' => 'Blog']);
 
         $this->assertTrue(is_file($this->modulePath . '/Http/Requests/CreateBlogPostRequest.php'));
+        $this->assertSame(0, $code);
     }
 
     /** @test */
     public function it_generated_correct_file_with_content()
     {
-        $this->artisan('module:make-request', ['name' => 'CreateBlogPostRequest', 'module' => 'Blog']);
+        $code = $this->artisan('module:make-request', ['name' => 'CreateBlogPostRequest', 'module' => 'Blog']);
 
         $file = $this->finder->get($this->modulePath . '/Http/Requests/CreateBlogPostRequest.php');
 
         $this->assertMatchesSnapshot($file);
+        $this->assertSame(0, $code);
     }
 
     /** @test */
@@ -55,11 +57,12 @@ class RequestMakeCommandTest extends BaseTestCase
     {
         $this->app['config']->set('modules.paths.generator.request.path', 'SuperRequests');
 
-        $this->artisan('module:make-request', ['name' => 'CreateBlogPostRequest', 'module' => 'Blog']);
+        $code = $this->artisan('module:make-request', ['name' => 'CreateBlogPostRequest', 'module' => 'Blog']);
 
         $file = $this->finder->get($this->modulePath . '/SuperRequests/CreateBlogPostRequest.php');
 
         $this->assertMatchesSnapshot($file);
+        $this->assertSame(0, $code);
     }
 
     /** @test */
@@ -67,10 +70,11 @@ class RequestMakeCommandTest extends BaseTestCase
     {
         $this->app['config']->set('modules.paths.generator.request.namespace', 'SuperRequests');
 
-        $this->artisan('module:make-request', ['name' => 'CreateBlogPostRequest', 'module' => 'Blog']);
+        $code = $this->artisan('module:make-request', ['name' => 'CreateBlogPostRequest', 'module' => 'Blog']);
 
         $file = $this->finder->get($this->modulePath . '/Http/Requests/CreateBlogPostRequest.php');
 
         $this->assertMatchesSnapshot($file);
+        $this->assertSame(0, $code);
     }
 }