+++ tests/Commands/__snapshots__/MigrationMakeCommandTest__it_generates_foreign_key_constraints__1.php
@@ -14,7 +14,7 @@ class CreatePostsTable extends Migration
     public function up()
     {
         Schema::create(\'posts\', function (Blueprint $table) {
-            $table->bigIncrements(\'id\');
+            $table->id();
 			$table->integer(\'user_id\')->unsigned();
 			$table->foreign(\'user_id\')->references(\'id\')->on(\'users\');