+++ tests/Commands/__snapshots__/ModuleMakeCommandTest__it_generates_correct_composerjson_file__1.php
@@ -9,9 +9,7 @@
     ],
     "extra": {
         "laravel": {
-            "providers": [
-                "Modules\\\Blog\\\Providers\\\BlogServiceProvider"
-            ],
+            "providers": [],
             "aliases": {
 
             }